+++ SE_GO/src/main/java/model/gameStatistics/GameStatistics.java
@@ -3,7 +3,7 @@ package model.gameStatistics;
 import singleComponents.EndgameColors;
 import singleComponents.StoneColor;
 
-public record GameStatistics(int blackPoints, int whitePoints, int blackMoves, int whiteMoves, int capturedBlackStones, int capturedWhiteStones, StoneColor winner, EndgameColors[][] endBoard) {
+public record GameStatistics(double blackPoints, double whitePoints, int blackMoves, int whiteMoves, int capturedBlackStones, int capturedWhiteStones, StoneColor winner, EndgameColors[][] endBoard) {
     @Override
     public String toString() {
         StringBuilder out = new StringBuilder("GameStats\n----------------------------------\n");
